requirements.txt
requests
semver
python-jenkins